+++ src_docs/components/input-row/input-row-component-view.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import { FormControl } from 'react-bootstrap';
-
+import variables from '@opuscapita/oc-cm-common-styles/styles/_app.scss';
 // App imports
-import { Layout, Content } from '../../../src/';
+import { Layout, Content, Primitive } from '../../../src/';
 
 export default class InputRowComponentView extends React.PureComponent {
   onChange = () => {
@@ -53,7 +52,7 @@ export default class InputRowComponentView extends React.PureComponent {
         <Content.Row>
           {this.getContent().map(section => (
             <Content.Column key={section.id}>
-              <Content.Card>
+              <Content.Card id={section.id}>
                 {section.fields.map((row) => {
                   if (row.readOnly) {
                     return <Content.ReadonlyRow key={row.id} label={row.label} value={row.value} />;
@@ -66,7 +65,7 @@ export default class InputRowComponentView extends React.PureComponent {
                       error={row.error}
                       required={row.required}
                     >
-                      <FormControl value={row.value} onChange={this.onChange} />
+                      <Primitive.Input value={row.value} onChange={this.onChange} />
                     </Content.InputRow>
                   );
                 })}